refactor(deps): update eslint-plugin-react-refresh to 0.5.3
Bump eslint-plugin-react-refresh from 0.4.26 to 0.5.3. Its
only-export-components rule is now stricter and flags object
literal exports alongside components.

Move the SETUP_SNAPSHOT constant out of App.tsx into a dedicated
src/constants/snapshot.ts module so App.tsx only exports its
component (which also restores Fast Refresh for that file), and
update the two importing files accordingly.

Verified with lint, build, test (166 passed), and format:check.
