+++ src/App.tsx
@@ -35,21 +35,6 @@ function isCacheAliveForTtlPolling(
 }
 
 /**
- * Constants for snapshot configuration
- */
-export const SETUP_SNAPSHOT = {
-  LIMIT: {
-    MAX: 10_000,
-    MIN: 0,
-    DEFAULT: 10,
-  },
-  OFFSET: {
-    MIN: 0,
-    DEFAULT: 0,
-  },
-} as const;
-
-/**
  * PromidasInfoSection - Display link to PROMIDAS documentation
  */
 function PromidasInfoSection() {
